% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Send request to data API /find endpoint}
\usage{
aggregate(collection, pipeline)
}
\arguments{
\item{collection}{mongekyo collection object}

\item{pipeline}{character, aggregation pipeline to subset/summarize documents
TODO: testing, input validation, output validation}
}
\description{
Send request to data API /find endpoint
}
