% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{asDocument}
\alias{asDocument}
\title{Converts data.frame to document format}
\usage{
asDocument(data)
}
\arguments{
\item{data}{data.frame to be converted to a array of documents in BSON format}
}
\description{
Converts data.frame to document format
}
