% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{deleteMany}
\alias{deleteMany}
\title{Send request to data API /insertMany endpoint
Inserts a one or many documents into a collection}
\usage{
deleteMany(collection, filter)
}
\arguments{
\item{collection}{collection object}

\item{filter}{value(s) by which to identify documents to be deleted
TODO: input validation}
}
\description{
Send request to data API /insertMany endpoint
Inserts a one or many documents into a collection
}
