% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{deleteOne}
\alias{deleteOne}
\title{Send request to data API /insertOne endpoint
Inserts a single document into collection}
\usage{
deleteOne(collection, filter)
}
\arguments{
\item{collection}{collection object}

\item{filter}{value(s) by which to identify document to be deleted
TODO: input validation}
}
\description{
Send request to data API /insertOne endpoint
Inserts a single document into collection
}
