% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find}
\alias{find}
\title{Send request to data API /find endpoint}
\usage{
find(collection, filter)
}
\arguments{
\item{collection}{mongekyo collection object}

\item{filter}{character, filter to subset documents
TODO: testing, input validation, output validation}
}
\description{
Send request to data API /find endpoint
}
