% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findOne}
\alias{findOne}
\title{Send request to data API /findOne endpoint}
\usage{
findOne(collection, filter)
}
\arguments{
\item{collection}{mongekyo collection object}

\item{filter}{character, filter to subset documents
TODO: testing, input validation, output validation}
}
\description{
Send request to data API /findOne endpoint
}
