% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{formatAggregationValue}
\alias{formatAggregationValue}
\title{Internal helper function to probably format arguments to accumulator functions}
\usage{
formatAggregationValue(value)
}
\arguments{
\item{value}{incomming argument to accumulator function}
}
\description{
Internal helper function to probably format arguments to accumulator functions
}
