% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{formatValue}
\alias{formatValue}
\title{Internal function to format vlaues according to their type}
\usage{
formatValue(value)
}
\arguments{
\item{value}{value to assign to a given field}
}
\value{
value formatted in BSON according to its data type
}
\description{
Internal function to format vlaues according to their type
}
