% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollection.R
\name{getCollection}
\alias{getCollection}
\title{Create a collection object
This function takes all base level information to access api and consolodates
it into a single object, which can then be passed to other exported functions to
create an API call}
\usage{
getCollection(url, cluster, database, collection, apiKey)
}
\arguments{
\item{url}{Base URL to mongoDB data api, can be copied from MongoDB Data API
page.}

\item{cluster}{cluster name}

\item{database}{database name}

\item{collection}{collection name}

\item{apiKey}{API key obtained from mongodb.  Recommend to store this in .Renviron file.

TODO: Add a bit of code which checks the connection to the collection}
}
\description{
Create a collection object
This function takes all base level information to access api and consolodates
it into a single object, which can then be passed to other exported functions to
create an API call
}
