% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{groupStage}
\alias{groupStage}
\title{Create a $group aggration stage}
\usage{
groupStage(id = NULL, accumulators = NULL)
}
\arguments{
\item{id}{character string, contains a field or expression to be used as the group key.}

\item{accumulators}{named list, list names will be the name of fields created within the output
document, list values are character strings containing accumulator expressions.}
}
\description{
Create a $group aggration stage
}
