% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insertOne}
\alias{insertOne}
\title{Send request to data API /insertOne endpoint
Inserts a single document into collection}
\usage{
insertOne(collection, document)
}
\arguments{
\item{collection}{collection object}

\item{document}{document in BSON format
TODO: input validation, add check if document already contains _id field}
}
\description{
Send request to data API /insertOne endpoint
Inserts a single document into collection
}
