% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{lookupStage}
\alias{lookupStage}
\title{Create a $lookup aggregation stage}
\usage{
lookupStage(from, localField, foreignField, as)
}
\arguments{
\item{from}{Specifies the collection in the same database to perform the join with}

\item{localField}{Specifies the field from the documents input to the $lookup stage}

\item{foreignField}{Specifies the field from the documents in the from collection}

\item{as}{Specifies the name of the new array field to add to the input documents

TODO implement optional parameters, see issue #5}
}
\description{
Create a $lookup aggregation stage
}
\examples{

# Create a lookup stage to get movie documents from comment document
lookupStage(from = "movies", localField = "movie_id",
            foreignField = "_id", as = "movieDocs")
  
# Use lookup stage in pipeline 
pipeline <- list(mongEq("Jojen Reed", "name") \%>\% matchStage(),
                 lookup = lookupStage(from = "movies", localField = "movie_id",
                                     foreignField = "_id", as = "movieDocs")) \%>\% 
  pipeline()
}
