% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{matchStage}
\alias{matchStage}
\title{Create a $match aggregation stage}
\usage{
matchStage(expression)
}
\arguments{
\item{expression}{a BSON expression which provides query conditions
(identical to an expression passed to the filter endpoint) by which to subset
the data.}
}
\description{
Create a $match aggregation stage
}
