% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{mongAnd}
\alias{mongAnd}
\title{Creates BSON formatted AND of two or more clauses
TODO: Input validation}
\usage{
mongAnd(clauses)
}
\description{
Creates BSON formatted AND of two or more clauses
TODO: Input validation
}
