% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{mongArray}
\alias{mongArray}
\title{Formats a vector as a BSON array}
\usage{
mongArray(values)
}
\arguments{
\item{values}{a vector which will be formatted based on its type, and formatted
as a BSON array}
}
\description{
Formats a vector as a BSON array
}
