% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{mongEq}
\alias{mongEq}
\title{Creates BSON formatted key-value pair}
\usage{
mongEq(value, field, kvPairs = NULL)
}
\arguments{
\item{value}{value}

\item{field}{field}
}
\description{
Creates BSON formatted key-value pair
}
