% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{mongOr}
\alias{mongOr}
\title{Creates BSON formatted OR of two or more expressions
TODO: Input validation}
\usage{
mongOr(expressions)
}
\description{
Creates BSON formatted OR of two or more expressions
TODO: Input validation
}
