% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{mongSum}
\alias{mongSum}
\title{Create expression with $sum accumulator}
\usage{
mongSum(expression)
}
\arguments{
\item{expression}{character string containing an expression or field which
evaluates as numeric.}
}
\description{
Create expression with $sum accumulator
}
\examples{
# sum the HOURS field
mongSum("HOURS")
# create count
mongSum(1)

}
