% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{pipeline}
\alias{pipeline}
\title{Create an aggregation pipeline}
\usage{
pipeline(stages)
}
\arguments{
\item{stages}{list of stages or character string containing stages.
If list, each element should contain a stage encased in curly braces.  If character
string, it should be a correctly formatted BSON array of stages
(use mongArray to convert an expression to BSON array).

TODO: Write code examples once more stages are written}
}
\description{
Create an aggregation pipeline
}
