% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{projectStage}
\alias{projectStage}
\title{Create a $project stage}
\usage{
projectStage(kvPairs)
}
\arguments{
\item{kvPairs}{named list, name provides name of field to be created, value
contains expression to generate the field's value}
}
\description{
Create a $project stage
}
\examples{
projStage <- projectStage(list(`title` = 1, `lead` = '{"$arrayElemAt": ["$cast", 0]}'))

\dontrun{
pipeline <- list(projectStage(list(`title` = 1, 
                                   `lead` = '{"$arrayElemAt": ["$cast", 0]}'))) \%>\% 
pipeline()
resp <- aggregate(movies, pipeline)
}
}
