% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateOne}
\alias{updateOne}
\title{Send request to data API /updateOne endpoint}
\usage{
updateOne(collection, filter, update)
}
\arguments{
\item{collection}{mongekyo collection object}

\item{filter}{character, filter to subset documents
TODO: testing, input validation, output validation}
}
\description{
Send request to data API /updateOne endpoint
}
