% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{kvCombine}
\alias{kvCombine}
\title{Internal function to combine fields and vlaues}
\usage{
kvCombine(value, field)
}
\arguments{
\item{value}{value to assign to a given field}

\item{field}{field to which value will be assigned}
}
\value{
string in BSON equal foramt with no surrounding brackets
}
\description{
Internal function to combine fields and vlaues
}
