% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{mongEq}
\alias{mongEq}
\title{Creates BSON formatted key-value pair}
\usage{
mongEq(value, field, kvPairs = NULL)
}
\arguments{
\item{value}{value}

\item{field}{field}

\item{kvPairs}{named list of key value pairs to be used if working with multiple fields and values.  value and field parameter will be ignored if this parameter is used.}
}
\description{
Creates BSON formatted key-value pair
}
